RingBuffer: make Generator advance the read pointer.
RingBuffer doesn't conform to Collection, so there's actually no reason
to ensure the Generator is multi-pass friendly.

The reference to the buffer is currently strong just because it should
live at least as long as the generator does, and the buffer doesn't
have a reference to the generator, so it's fine.